requirements.txt
streamlit
google-generativeai
audio-recorder-streamlit
SpeechRecognition
pydub
